# Dynamic Open Model Zoo as Python module