--- a/tools/downloader/test.py
+++ b/tools/downloader/test.py
@@ -3,7 +3,7 @@ import os
 #
 # Test paths from MO args
 #
-from topologies import ssd300
+from topologies.public import ssd300
 
 t = ssd300()
 xmlPath, binPath = t.getIR()
@@ -16,3 +16,10 @@ for path, name in zip ([t.model, t.config, xmlPath, binPath, xmlPath_fp16, binPa
                         'ssd300.xml', 'ssd300.bin']):
     assert(os.path.exists(path)), path
     assert(os.path.basename(path) == name), name
+
+#
+# Check aliases
+#
+from topologies.intel import vehicle_license_plate_detection_barrier_0106
+
+t = vehicle_license_plate_detection_barrier_0106()