--- a/tools/downloader/topologies.py
+++ b/tools/downloader/topologies.py
@@ -4,91 +4,86 @@ import re
 import yaml
 import types
 
+from pathlib import Path
+
 import common
 
 if 'INTEL_OPENVINO_DIR' in os.environ:  # As a part of OpenVINO
-    _loc = os.path.join(os.environ['INTEL_OPENVINO_DIR'], 'deployment_tools', 'open_model_zoo')
-    _modelsDir = os.path.join(_loc, 'models')
-    _downloader = os.path.join(_loc, 'tools', 'downloader', 'downloader.py')
+    _loc = Path(os.environ['INTEL_OPENVINO_DIR']) / 'deployment_tools' / 'open_model_zoo'
+    _models_dir = _loc / 'models'
+    _downloader = _loc / 'tools' / 'downloader' / 'downloader.py'
 else:  # Standalone
-    _loc = os.path.dirname(__file__)
-    _modelsDir = os.path.join(_loc, '..', '..', 'models')
-    _downloader = os.path.join(_loc, 'downloader.py')
+    _loc = Path(__file__).parent
+    _models_dir = _loc / '../../models'
+    _downloader = _loc / 'downloader.py'
 
-_precisionMarker = '$precision'
-_cache = os.path.abspath(os.getenv('OPENCV_OPEN_MODEL_ZOO_CACHE_DIR', 'models'))
+_precision_marker = '$precision'
+_cache = Path(os.getenv('OPENCV_OPEN_MODEL_ZOO_CACHE_DIR', 'models')).resolve()
 
 # Parse topologies.
-for _moduleName in os.listdir(_modelsDir):
-    _modelsSubDir = os.path.join(_modelsDir, _moduleName)
-    if not os.path.isdir(_modelsSubDir):
+for _module_name in os.listdir(_models_dir):
+    _models_subdir = _models_dir / _module_name
+    if not _models_subdir.is_dir():
         continue
 
-    _module = types.ModuleType(_moduleName)
-    sys.modules['topologies.' + _moduleName] = _module
-    globals()[_moduleName] = _module
+    _module = types.ModuleType(_module_name)
+    sys.modules['topologies.' + _module_name] = _module
+    globals()[_module_name] = _module
     _module.__dict__['sys'] = sys
     _module.__dict__['os'] = os
 
-    # List of models which have versions (mostly DLDT models). For such kind of models
-    # with the highest version we will add short names. In example, for
-    # "face_detection_retail_0004" and "face_detection_retail_0005" method with name
-    # "face_detection_retail" returns result of "face_detection_retail_0005".
-    _versionedNames = {}
-
-    for _topologyName in os.listdir(_modelsSubDir):
-        _config = os.path.join(_modelsSubDir, _topologyName, 'model.yml')
-        if not os.path.exists(_config):
+    for _topology_name in os.listdir(_models_subdir):
+        _config = _models_subdir / _topology_name / 'model.yml'
+        if not _config.exists():
             continue
 
         with open(_config, 'rt') as _f:
             _topology = yaml.safe_load(_f)
-            _name = _topologyName.replace('-', '_').replace('.', '_')
+            _name = _topology_name.replace('-', '_').replace('.', '_')
             _description = _topology['description']
             _license = _topology['license']
             _files = _topology['files']
             _framework = _topology['framework']
-            _taskType = _topology['task_type']
-            _configPath = ''
-            _modelPath = ''
-            _downloadDir = os.path.join(_cache, _moduleName, _topologyName)
-            _moArgs = {}
+            _task_type = _topology['task_type']
+            _config_path = ''
+            _model_path = ''
+            _download_dir = _cache / _module_name / _topology_name
+            _mo_args = {}
             if 'model_optimizer_args' in _topology:
                 for _arg in _topology['model_optimizer_args']:
                     _tokens = _arg.split('=')
-                    _moArgs[_tokens[0]] = _tokens[1] if len(_tokens) == 2 else None
+                    _mo_args[_tokens[0]] = _tokens[1] if len(_tokens) == 2 else None
 
 
             assert(len(_files) > 0)
             if len(_files) > 2:
                 assert(_framework == 'dldt'), ('Unexpected framework type: ' + _framework)
                 # Get only basename and add precision marker
-                _configPath = _files[0]['name']
-                _modelPath = _files[1]['name']
+                _config_path = _files[0]['name']
+                _model_path = _files[1]['name']
 
-                _pos = _configPath.find('/')
-                if _configPath[:_pos] in common.KNOWN_PRECISIONS:
-                    _configPath = _precisionMarker + _configPath[_pos:]
+                _pos = _config_path.find('/')
+                if _config_path[:_pos] in common.KNOWN_PRECISIONS:
+                    _config_path = _precision_marker + _config_path[_pos:]
 
-                _pos = _modelPath.find('/')
-                if _modelPath[:_pos] in common.KNOWN_PRECISIONS:
-                    _modelPath = _precisionMarker + _modelPath[_pos:]
+                _pos = _model_path.find('/')
+                if _model_path[:_pos] in common.KNOWN_PRECISIONS:
+                    _model_path = _precision_marker + _model_path[_pos:]
 
             else:
-                _configPath = _files[0]['name']
-                _modelPath = _files[-1]['name']
+                _config_path = _files[0]['name']
+                _model_path = _files[-1]['name']
 
             # To manage origin files location from archives
-            if '--input_model' in _moArgs:
-                _modelPath = _moArgs['--input_model'].replace('$dl_dir/', '')
-            if '--input_proto' in _moArgs:
-                _configPath = _moArgs['--input_proto'].replace('$dl_dir/', '')
+            if '--input_model' in _mo_args:
+                _model_path = _mo_args['--input_model'].replace('$dl_dir/', '')
+            if '--input_proto' in _mo_args:
+                _config_path = _mo_args['--input_proto'].replace('$dl_dir/', '')
 
-            if _configPath:
-                _configPath = os.path.realpath(os.path.join(_downloadDir, _configPath))
-            if _modelPath:
-                _modelPath = os.path.realpath(os.path.join(_downloadDir, _modelPath))
-
+            if _config_path:
+                _config_path = (_download_dir / _config_path).resolve()
+            if _model_path:
+                _model_path = (_download_dir / _model_path).resolve()
 
             exec("""
 class {className}:
@@ -213,21 +208,11 @@ class {className}:
         return IENetwork(xmlPath, binPath)
 
             """.format(className=_name, description=_description, license=_license,
-                       name=_topologyName, outdir=_cache,
-                       config=_configPath, model=_modelPath,
-                       framework=_framework, downloader=_downloader, module=_moduleName,
-                       precisionMarker=_precisionMarker, task_type=_taskType,
-                       mo_args=_moArgs),
+                       name=_topology_name, outdir=str(_cache).replace('\\', '\\\\'),
+                       config=str(_config_path).replace('\\', '\\\\'),
+                       model=str(_model_path).replace('\\', '\\\\'),
+                       framework=_framework, module=_module_name,
+                       downloader=str(_downloader).replace('\\', '\\\\'),
+                       precisionMarker=_precision_marker, task_type=_task_type,
+                       mo_args=_mo_args),
             _module.__dict__)
-
-            # Extract version and add an alias.
-            _matches = re.search('(.+)_(\d{4})$', _name)
-            if _matches:
-                _shortName = _matches.group(1)
-                _version = _matches.group(2)
-                if not _shortName in _versionedNames or int(_version) > int(_versionedNames[_shortName]):
-                    _versionedNames[_shortName] = _version
-                    exec("""
-{alias} = {className}
-                    """.format(alias=_shortName, className=_name),
-                    _module.__dict__)