**WIP**

an alternative solution for https://github.com/opencv/open_model_zoo/pull/272 with limited usage (no demos and no C++ but much more simpler).

- [ ] Decide location of `topologies.py` file to make it possible:
  ```python
  from openvino.open_model_zoo.topologies import mobilenet_ssd
  ```
- [x] Relative paths: to models and to `downloader.py` and `converter.py` scripts
- [ ] Add an ability to append / override MO arguments (in example, to vary precision)
- [x] Aliases for DLDT networks versions (like in https://github.com/opencv/open_model_zoo/pull/272)
- [x] Tests
- [x] OpenCV conversion (with mean and scales setting)
- [ ] Try to emit `requests` package dependency (non default)

## Current state:

* Documentation

```python
from topologies import mobilenet_ssd
help(mobilenet_ssd)
```
```
Help on class mobilenet_ssd in module topologies:

class mobilenet_ssd(Topology)
 |      The `mobilenet-ssd` model is a Single-Shot multibox Detection (SSD) network intended to perfo
rm object detection. This model is implemented using the Caffe\* framework. For details about this mo
del, check out the repository <https://github.com/chuanqi305/MobileNet-SSD>.
 |  The model input is a blob that consists of a single image of 1x3x300x300 in BGR order, also like 
the `densenet-121` model. The BGR mean values need to be subtracted as follows: [127.5, 127.5, 127.5]
 before passing the image blob into the network. In addition, values must be divided by 0.007843.
 |  The model output is a typical vector containing the tracked object data, as previously described.
 |  
 |      License: https://raw.githubusercontent.com/chuanqi305/MobileNet-SSD/master/LICENSE
 |  
 |      Attributes
 |      ----------
 |          config : str
 |              Absolute path to text file with network configuration.
 |          model : str
 |              Absolute path to weights file of network.
 |          framework : str
 |              Name of the framework in which format network is represented.
 |  
 |  Method resolution order:
 |      mobilenet_ssd
 |      Topology
 |      builtins.object
 |  
 |  Methods defined here:
 |  
 |  __init__(self, precision='FP32')
 |      Initialize self.  See help(type(self)) for accurate signature.
 |  
 |  ----------------------------------------------------------------------
 |  Methods inherited from Topology:
 |  
 |  getIENetwork(self)
 |      Creates openvino.inference_engine.IENetwork instance.
 |      Model Optimizer is lauched to create OpenVINO Intermediate Representation (IR).
 |  
 |  getIR(self, precision='FP32')
 |      Creates OpenVINO Intermediate Representation (IR) network format.
 |      
 |      Returns
 |      -------
 |          xmlPath
 |              Path to generated or downloaded .xml file
 |          binPath
 |              Path to generated or downloaded .bin file
 |  
 |  getOCVModel(self, useIR=True)
 |      Creates OpenCV's cv::dnn::Model from origin network. Depends on network type,
 |      returns cv.dnn_DetectionModel, cv.dnn_ClassificationModel, cv.dnn_SegmentationModel
 |      or cv.dnn_Model if not specified.
 |      
 |      Preprocessing parameters are set.
 |  
 |  ----------------------------------------------------------------------
 |  Data descriptors inherited from Topology:
 |  
 |  __dict__
 |      dictionary for instance variables (if defined)
 |  
 |  __weakref__
 |      list of weak references to the object (if defined)
```

* Downloading
```python
from topologies import mobilenet_ssd
ssd = mobilenet_ssd()
```
```
###############|| Downloading models ||###############

========= Downloading /home/dkurtaev/open_model_zoo/tools/downloader/models/public/mobilenet-ssd/mobilenet-ssd.prototxt
... 100%, 28 KB, 47565 KB/s, 0 seconds passed

========= Downloading /home/dkurtaev/open_model_zoo/tools/downloader/models/public/mobilenet-ssd/mobilenet-ssd.caffemodel
... 100%, 22605 KB, 6388 KB/s, 3 seconds passed


###############|| Post processing ||###############
```

### Usage example
* With OpenCV
```python
import cv2 as cv

from topologies import mobilenet_ssd

ssd = mobilenet_ssd()
net = ssd.getOCVModel()

frame = cv.imread('example.jpg')

classIds, confidences, boxes = net.detect(frame, confThreshold=0.5)
for box in boxes:
    cv.rectangle(frame, box, (0, 255, 0))

cv.imshow('out', frame)
cv.waitKey()
```
![out](https://user-images.githubusercontent.com/25801568/63178190-30cb6c80-c052-11e9-846a-86b323c8c4d3.png)

* With Inference Engine
```python
from topologies import mobilenet_ssd
from openvino.inference_engine import IENetwork

ssd = mobilenet_ssd()

xmlPath, binPath = ssd.getIR()
net = IENetwork(xmlPath, binPath)
```