@dkurt 2019-08-18
@IRDonch, @vladimir-dudnik, Please review this PR. This is Python only tool to download models using `downloader.py` and convert them to IRs by `converter.py`.

Open question is location of this script in OpenVINO distribution. I want to see something like
```python
import openvino.open_model_zoo.topologies.public
import openvino.open_model_zoo.topologies.intel
```