@IRDonch 2019-08-29
> Is there any chance that we can add this module to 2019R3?

Honestly, the chances are pretty slim. I still need to review a lot of the code, plus there needs to be documentation and (preferably) CI integration.