@dkurt 2019-12-14
Removed tests for now. We can enable them after merge of this PR.