@dkurt 2019-12-18
@IRDonch, @vladimir-dudnik, @ngaloppo, may I ask you to review this patch one more time? If there are no critical issues which we cannot fix further without breaking compatibility, can we merge it now? Tests and integration questions can be solved separately due they depend on different repositories.