@ngaloppo 2019-12-18
First, let me comment on the value of a library that can load models from the model zoo, _in addition to a program to do so_. In particular, this is useful for interactive experimentation notebooks (e.g. Jupyter), so that you don't need to go through shell execution to call `downloader.py` (and potentially have to set up the shell to have the program dependencies set up). The requirements for the program and the library are different. The program will download and store serialized versions of the models to the filesystem, immediately. The behavior and features of the library is different - so let's align on those first. I have taken some inspiration from `torch.utils.model_zoo` and `tf.hub`. 

### Features

I don't think the current approach (which mimics `torchvision`) is maintainable. It requires creating classes for each new model in the model zoo. So please allow me to suggest a slight difference to the approach. 

#### Minimal features

- Get available models in [open\_model\_zoo](https://github.com/opencv/open_model_zoo), return a list of model names
- Load a model by model name (e.g. `'se-resnext-50'`), return an instance of `IENetwork`

That's it. All else is transparent to the user. Users should not have to care about where or how models are stored/cached. This approach implies that the model is downloaded immediately (if not already cached), unless we make significant changes to how IENetwork works, I **think**. The user can convert to serialized version by using `IENetwork.serialize()`.

#### Optional features

- Load a model by URL pointing to a .tgz file (containing any supported serialized version)
- Get references to the cached serialized (OpenVINO) IR of the model.
- Some way to get a `cv::dnn::Model` out of this library. Is there a way to convert `IENetwork` instances to `cv::dnn::Model` instances? Alternatively, if we implement the feature above, we could go through the `cv::dnn::readNet()` path. What 

#### API

These are the minimal API's I'd like to see.  Let's iterate over the API, and fine-tune the behavior here.

- `def get_model_names() -> List[str]`
- `def get_model(name: str, progress: bool = True, cache_location: Optional[str]) -> openvino.IENetwork`

Basic usage: 

```
import openvino.open_model_zoo as omz

name = 'resnet-50'
if name in omz.get_model_names():
	model = omz.load_model(name)
...
```

#### Notes

**Download only vs. load model**: I could be convinced to make it even simpler: 

```
import openvino.open_model_zoo as omz
name = 'resnet-50'
xml, bin = omz.download_model(name)
model = IENetwork.from_ir(xml, bin)
```

The API spec would be `download_model(name:str, progress: bool = True, cache_location: Optional[str]) -> Tuple[str, str]` returning the paths of the cached `.xml,.bin` files. 

This also solves the problem of wanting a `cv::dnn::Model`, and not a `ie_api.IENetwork`.

**Cached location**: I like the approach in `torch.utils.model_zoo`:

> If the object is already present in model_dir, it’s deserialized and returned. The default value of model_dir is $TORCH_HOME/checkpoints where environment variable $TORCH_HOME defaults to $XDG_CACHE_HOME/torch. $XDG_CACHE_HOME follows the X Design Group specification of the Linux filesytem layout, with a default value ~/.cache if not set.

Replace `$TORCH_HOME` by `$OPEN_MODEL_ZOO_HOME`.