@IRDonch 2019-12-19
I'm certainly all for starting with a minimal viable API.

I'd like to suggest a modification, though. I think the usage should look like this:

```python
import openvino.open_model_zoo as omz

with omz.Zoo() as zoo:
    name = 'resnet-50'
    assert name in zoo.model_names:
	model = zoo.load_model(name)
```

The loading of information about the modules would be encapsulated in `Zoo.__init__`, so that the importing of the module by itself doesn't trigger reading of config files / launching programs. And `Zoo` is a context manager, so that it can remove any temporary files it creates by the end of the program.