@IRDonch 2019-12-19
> First, let me comment on the value of a library that can load models from the model zoo, in addition to a program to do so. In particular, this is useful for interactive experimentation notebooks (e.g. Jupyter), so that you don't need to go through shell execution to call downloader.py (and potentially have to set up the shell to have the program dependencies set up).

FWIW, this could potentially work for models that are published in the IR format, but for other models you have to run Model Optimizer (and potentially other scripts) to convert them, so the ability to execute external programs is required anyway.