@ngaloppo 2019-12-20
> FWIW, this could potentially work for models that are published in the IR format, but for other models you have to run Model Optimizer (and potentially other scripts) to convert them, so the ability to execute external programs is required anyway.

I think that support for IR format models only is acceptable for the minimum viable feature set of the first version. What's the percentage of OMZ models that has IR format published?