@dkurt 2020-07-24
see https://github.com/openvinotoolkit/open_model_zoo/pull/1352